# src/api/openapi/v3.yaml
{
  "openapi": "3.1.0",
  "info": {"title": "EdgeWARN Unified API", "version": "3.0.0", "description": "The versioned, read-only EdgeWARN and EWMRS public API."},
  "servers": [{"url": "/"}],
  "paths": {
    "/api/v3": {"get": {"operationId": "getApiDiscovery", "responses": {"200": {"description": "Capability document", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResourceEnvelope"}}}}}}},
    "/api/v3/openapi.json": {"get": {"operationId": "getOpenApi", "responses": {"200": {"description": "OpenAPI document"}}}},
    "/health/live": {"get": {"operationId": "getLiveness", "responses": {"200": {"description": "Live"}}}},
    "/health/ready": {"get": {"operationId": "getReadiness", "responses": {"200": {"description": "Ready"}, "503": {"description": "Producer roots unavailable"}}}},
    "/api/v3/cells": {"get": {"operationId": "listCells", "responses": {"200": {"description": "Cell collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/cells/{cellId}": {"get": {"operationId": "getCell", "parameters": [{"$ref": "#/components/parameters/cellId"}], "responses": {"200": {"description": "Cell"}, "404": {"$ref": "#/components/responses/Problem"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/storm-snapshots": {"get": {"operationId": "listStormSnapshots", "responses": {"200": {"description": "Snapshot collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/storm-snapshots/{timestamp}": {"get": {"operationId": "getStormSnapshot", "parameters": [{"$ref": "#/components/parameters/timestamp"}], "responses": {"200": {"description": "Storm snapshot"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/alert-snapshots": {"get": {"operationId": "listAlertSnapshots", "parameters": [{"$ref": "#/components/parameters/source"}], "responses": {"200": {"description": "Alert snapshot collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/alert-snapshots/{timestamp}": {"get": {"operationId": "getAlertSnapshot", "parameters": [{"$ref": "#/components/parameters/timestamp"}, {"$ref": "#/components/parameters/source"}], "responses": {"200": {"description": "Alert snapshot"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/alerts/{alertId}": {"get": {"operationId": "getAlert", "parameters": [{"$ref": "#/components/parameters/alertId"}, {"$ref": "#/components/parameters/source"}], "responses": {"200": {"description": "Alert"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/observations/metar": {"get": {"operationId": "listMetarObservationHours", "responses": {"200": {"description": "METAR observation hour collection"}}}},
    "/api/v3/observations/metar/{timestamp}": {"get": {"operationId": "getMetarObservations", "parameters": [{"$ref": "#/components/parameters/timestamp"}], "responses": {"200": {"description": "METAR observations"}}}},
    "/api/v3/render-products": {"get": {"operationId": "listRenderProducts", "responses": {"200": {"description": "Binary-chunk render product collection; scalar visualization styles are client-owned"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/render-products/{productId}": {"get": {"operationId": "getRenderProduct", "parameters": [{"$ref": "#/components/parameters/productId"}], "responses": {"200": {"description": "Render product using the binary_chunks representation"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/render-products/{productId}/snapshots": {"get": {"operationId": "listRenderSnapshots", "parameters": [{"$ref": "#/components/parameters/productId"}], "responses": {"200": {"description": "Render snapshot collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/render-products/{productId}/snapshots/{timestamp}/chunks": {"get": {"operationId": "listRenderChunks", "parameters": [{"$ref": "#/components/parameters/productId"}, {"$ref": "#/components/parameters/timestamp"}], "responses": {"200": {"description": "Schema-versioned float16 chunk coordinate collection"}, "503": {"description": "Required EWMRS service is not active (SERVICE_NOT_ENABLED), or the requested render artifact is incomplete or invalid (INVALID_ARTIFACT)", "content": {"application/problem+json": {}}}}}},
    "/api/v3/render-products/{productId}/snapshots/{timestamp}/chunks/{x}/{y}": {"get": {"operationId": "getRenderChunk", "parameters": [{"$ref": "#/components/parameters/productId"}, {"$ref": "#/components/parameters/timestamp"}, {"$ref": "#/components/parameters/x"}, {"$ref": "#/components/parameters/y"}], "responses": {"200": {"description": "float16 scalar or RGB value chunk (top-to-bottom rows; bottom-left grid origin; NaN no-data)", "headers": {"X-EWMRS-Format-Version": {"schema": {"type": "string", "example": "2"}}, "X-Data-Type": {"schema": {"type": "string", "example": "float16"}}, "X-Channel-Count": {"schema": {"type": "string", "example": "1"}}, "X-Chunk-Width": {"schema": {"type": "string", "example": "350"}}, "X-Chunk-Height": {"schema": {"type": "string", "example": "350"}}}, "content": {"application/octet-stream": {}}}, "503": {"description": "Required EWMRS service is not active (SERVICE_NOT_ENABLED), or the requested render artifact is incomplete or invalid (INVALID_ARTIFACT)", "content": {"application/problem+json": {}}}}}},
    "/api/v3/radar-sites": {"get": {"operationId": "listRadarSites", "responses": {"200": {"description": "Radar site collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/radar-sites/{siteId}/availability": {"get": {"operationId": "getRadarAvailability", "parameters": [{"$ref": "#/components/parameters/siteId"}], "responses": {"200": {"description": "Radar availability"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/radar-sites/{siteId}/scans/{timestamp}/elevations/{elevation}/products/{productId}": {"get": {"operationId": "getRadarField", "parameters": [{"$ref":"#/components/parameters/siteId"}, {"$ref":"#/components/parameters/timestamp"}, {"$ref":"#/components/parameters/elevation"}, {"$ref":"#/components/parameters/radarProductId"}], "responses": {"200": {"description": "Gzip-compressed polar field", "content": {"application/gzip": {}}}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layers": {"get": {"operationId": "listRapLayers", "responses": {"200": {"description": "RAP layer collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layers/{layerId}/snapshots": {"get": {"operationId": "listRapSnapshots", "parameters": [{"$ref":"#/components/parameters/layerId"}], "responses": {"200": {"description": "RAP snapshot collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layers/{layerId}/snapshots/{timestamp}/metadata": {"get": {"operationId": "getRapMetadata", "parameters": [{"$ref":"#/components/parameters/layerId"}, {"$ref":"#/components/parameters/timestamp"}], "responses": {"200": {"description": "RAP metadata"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layers/{layerId}/snapshots/{timestamp}/data": {"get": {"operationId": "getRapData", "parameters": [{"$ref":"#/components/parameters/layerId"}, {"$ref":"#/components/parameters/timestamp"}], "responses": {"200": {"description": "Uint16 data", "content": {"application/octet-stream": {}}}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/analyses/wpc/surface": {"get": {"operationId": "listWpcSurfaceAnalyses", "responses": {"200": {"description": "WPC surface analysis collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/analyses/wpc/surface/{timestamp}": {"get": {"operationId": "getWpcSurfaceAnalysis", "parameters": [{"$ref":"#/components/parameters/timestamp"}], "responses": {"200": {"description": "Surface analysis GeoJSON", "content": {"application/geo+json": {}}}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/modules": {"get": {"operationId": "listModules", "responses": {"200": {"description": "Modules declaring public routes", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModuleCollectionEnvelope"}}}}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/modules/{moduleId}": {"get": {"operationId": "getModule", "parameters": [{"$ref": "#/components/parameters/moduleId"}], "responses": {"200": {"description": "Module descriptor", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModuleEnvelope"}}}}, "400": {"description": "Malformed identifier (INVALID_PATH)", "$ref": "#/components/responses/Problem"}, "404": {"$ref": "#/components/responses/Problem"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/modules/{moduleId}/{routeId}": {"get": {"operationId": "getModuleRoute", "parameters": [{"$ref": "#/components/parameters/moduleId"}, {"$ref": "#/components/parameters/routeId"}], "responses": {"200": {"description": "Latest committed module JSON payload", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModuleRouteEnvelope"}}}}, "400": {"description": "Malformed identifier (INVALID_PATH)", "$ref": "#/components/responses/Problem"}, "404": {"$ref": "#/components/responses/Problem"}, "503": {"description": "EdgeWARN is disabled/stale (SERVICE_NOT_ENABLED), or the declared route has no committed representation (MODULE_ROUTE_UNAVAILABLE)", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}}}}
  },
  "components": {
    "schemas": {
      "ResourceEnvelope": {
        "type": "object",
        "required": ["data", "meta"],
        "additionalProperties": false,
        "properties": {
          "data": {"type": "object"},
          "meta": {"type": "object"}
        }
      },
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "instance", "requestId"],
        "properties": {
          "type": {"type": "string"},
          "title": {"type": "string"},
          "status": {"type": "integer", "minimum": 400, "maximum": 599},
          "detail": {"type": "string"},
          "instance": {"type": "string"},
          "requestId": {"type": "string"},
          "code": {"type": "string"}
        }
      },
      "ModuleRouteDescriptor": {"type": "object", "required": ["id", "description", "href", "available"], "additionalProperties": false, "properties": {"id": {"type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$"}, "description": {"type": "string", "maxLength": 256}, "href": {"type": "string"}, "available": {"type": "boolean"}}},
      "ModuleDescriptor": {"type": "object", "required": ["id", "name", "version", "href", "routes"], "additionalProperties": false, "properties": {"id": {"type": "string"}, "name": {"type": "string"}, "version": {"type": "string"}, "href": {"type": "string"}, "routes": {"type": "array", "items": {"$ref": "#/components/schemas/ModuleRouteDescriptor"}}}},
      "ModuleCollectionEnvelope": {"type": "object", "required": ["data", "meta"], "additionalProperties": false, "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/ModuleDescriptor"}}, "meta": {"type": "object"}}},
      "ModuleEnvelope": {"type": "object", "required": ["data", "meta"], "additionalProperties": false, "properties": {"data": {"$ref": "#/components/schemas/ModuleDescriptor"}, "meta": {"type": "object"}}},
      "ModuleRouteEnvelope": {"type": "object", "required": ["data", "meta"], "additionalProperties": false, "properties": {"data": {}, "meta": {"type": "object", "required": ["moduleId", "moduleVersion", "routeId", "cycleId", "publishedAt"], "properties": {"moduleId": {"type": "string"}, "moduleVersion": {"type": "string"}, "routeId": {"type": "string"}, "cycleId": {"type": "string"}, "publishedAt": {"type": "string", "format": "date-time"}}}}}
    },
    "parameters": {
      "cellId": {"name": "cellId", "in": "path", "required": true, "schema": {"type": "string"}},
      "alertId": {"name": "alertId", "in": "path", "required": true, "schema": {"type": "string"}},
      "productId": {"name": "productId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Za-z0-9_.-]+$"}},
      "siteId": {"name": "siteId", "in": "path", "required": true, "schema": {"type": "string"}},
      "layerId": {"name": "layerId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Za-z0-9_.-]+$"}},
      "elevation": {"name": "elevation", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^\\d{1,3}(?:\\.\\d{1,2})?$"}},
      "radarProductId": {"name": "productId", "in": "path", "required": true, "schema": {"type": "string", "enum": ["DBZH", "VRADH", "WRADH", "PHIDP", "CCORH", "RHOHV", "ZDR"]}},
      "timestamp": {"name": "timestamp", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^\\d{8}-\\d{6}$"}},
      "source": {"name": "source", "in": "query", "required": true, "schema": {"type": "string", "enum": ["official", "edgewarn"]}},
      "x": {"name": "x", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 0}},
      "y": {"name": "y", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 0}}
      ,"moduleId": {"name": "moduleId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$"}}
      ,"routeId": {"name": "routeId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$"}}
    },
    "responses": {"Problem": {"description": "RFC 9457 problem detail", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}, "ServiceNotEnabled": {"description": "Required service is not active (SERVICE_NOT_ENABLED; carries service, state, and lastSeen extension members)", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}}
  }
}